moving full to a ecma6, track-based viewer
