+++ demo/src/index.HTML
@@ -4,7 +4,6 @@
 		<link rel="stylesheet" href="app.css">
 	<body>
 		<p>Basic Demo</p>
-		<script src="bundle.js"></script>
 		<form id="theForm">
 			Genome : <input type="text" name="genome"><br>
 			Start : <input type="text" name="start"><br>
@@ -14,5 +13,6 @@
 		</form>
 		<svg className="viewer" id="viewer"/><br>
 		<svg className="viewer2" id="viewer2"/>
+		<script src="bundle.js"></script>
 	</body>
 </html>
